requirements.txt
undetected-chromedriver
selenium 
webdriver-manager
pandas
requests
PyPDF2
pytesseract
pdf2image
pillow